isUsernameExists moved to the UserDAO
